+++ src/matterbridge.ts
@@ -257,6 +257,9 @@ export class Matterbridge extends EventEmitter<MatterbridgeEvents> {
     if (this.server.isWorkerResponse(msg, msg.type)) {
       this.log.debug(`**Received broadcast response ${CYAN}${msg.type}${db} from ${CYAN}${msg.src}${db}: ${debugStringify(msg)}${db}`);
       switch (msg.type) {
+        case 'get_log_level':
+        case 'set_log_level':
+          break;
         default:
           this.log.debug(`Unknown broadcast response ${CYAN}${msg.type}${db} from ${CYAN}${msg.src}${db}`);
       }